+++ packages/next-common/components/overview/activeProposals/columns/common.jsx
